+++ src/client/index.ts
@@ -5,7 +5,6 @@ import events from '../events'
 import keys from '../keys'
 const fs = require('fs');
 
-export let blackcards: string[] = [];
 const client = new Client({
   intents: [
     GatewayIntentBits.Guilds,
@@ -13,21 +12,6 @@ const client = new Client({
   ],
 })
 
-fs.readFile('src/client/cah.json', 'utf8', (err: any, data: any) => {
-  if (err) {
-    console.error(err);
-    return;
-  }
-  const json = JSON.parse(data);
-  json.map((obj: any) => {
-    blackcards = blackcards.concat(obj.black);
-  })
-  blackcards = blackcards.filter((obj: any) => {
-    return obj.pick === 1
-  })
-  blackcards = blackcards.map((obj: any) => { return obj.text.replace(/_/g, '[BLANK]');;});
-  //console.log(blackcards)
-});
 
 registerEvents(client, events)
 
@@ -39,7 +23,7 @@ client.login(keys.clientToken).then(async () => {
       status: 'dnd',
     })
   })
-  .catch((err) => {
+  .catch((err: any) => {
     console.error('[Login Error]', err)
     process.exit(1)
   })
